Refactoring; group code by responsibility.
